+++ CHANGELOG.md
@@ -10,6 +10,15 @@ This project does not follow SemVer, since modules are independent of each other
 ### s3-private
 - Add default value for sse_algorithm to AES256. [#290](https://github.com/dbl-works/terraform/pull/290)
 
+### rds
+- Set nullable to false for engine_version. [#292](https://github.com/dbl-works/terraform/pull/292)
+
+### circleci-token-rotator
+- New module to rotate IAM user's AWS Access Keys on CircleCI. [#278](https://github.com/dbl-works/terraform/pull/278)
+
+### stack/app
+- Set default rds_engine_version to 16.
+
 ### ecs-deploy/service
 - Add vpc_name variables. [#287](https://github.com/dbl-works/terraform/pull/287)
 
@@ -18,6 +27,7 @@ This project does not follow SemVer, since modules are independent of each other
 
 ### ecs
 - Add enable_container_insights variables. [#282](https://github.com/dbl-works/terraform/pull/282)
+
 
 ## [v2023.12.03]
 ### ecs